+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 Version Master
 =======================================
+* Added a windows service wrapper class for using pulsar with windows services.
 * Removed the `worker` module.
 * Moved `http.rpc` module to `apps`.
 * Introduced context manager for `apps.tasks` to handle logs and exceptions.